Push working code
Code is currently working pretty well - we'll want to analyze the performance and look for any issues that may be glaringly obvious out of the plots. For the most part, cartesian wise, the truth values follow the waypoints well, and the kalman filter does a good job predicting it. The euler angles would be the place to investigate if there are any issues since they wouldn't be as obvious.